% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTransmittedRareVarCounts.R
\name{getTransmittedRareVarCounts}
\alias{getTransmittedRareVarCounts}
\title{Get counts of the numberof times each rare variants transmitted from parents to offspring in the data set
`getTransmittedRareVarCounts()` returns a data frame containing the count of rare variants =transmitted from the parents to the affected offspring. vcf, ped, cutoff = 0.01}
\usage{
getTransmittedRareVarCounts(vcf, ped, cutoff = 0.01)
}
\arguments{
\item{vcf}{vcf file}

\item{ped}{data frame containing pedigree information for the VCF}

\item{cutoff}{The MAF cutoff to determine which variants are "rare" (default is 0.01), for which transmission counts will calculated.}
}
\value{
results data frame containing the names of the SNVs that are considered rare, the number of times the rare variant was transmitted from parents to the affected offspring in the dataset, and the genomic position of the rare SNPs.
}
\description{
Get counts of the numberof times each rare variants transmitted from parents to offspring in the data set
`getTransmittedRareVarCounts()` returns a data frame containing the count of rare variants =transmitted from the parents to the affected offspring. vcf, ped, cutoff = 0.01
}
