% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV_TDT.R
\name{RV_TDT}
\alias{RV_TDT}
\title{Calculating RV-TDT statistic for user-defined windows}
\usage{
RV_TDT(plink.ped = NULL, vcf = NULL, vcf.ped = NULL, rv.tdt.dir,
  window.size = 0, window.type = "M", u = 0.2, param)
}
\arguments{
\item{window.size}{window size size of window}

\item{window.type}{type of window, either number of markers ("M") or width of kilobase interval ("K")}

\item{param}{parameters for RV-TDT}

\item{filepath.vcf}{absolute filepath of vcf}

\item{filepath.ped}{absolute filepath of vcf}
}
\value{
results data frame containing results from RV-TDT
}
\description{
Calculating RV-TDT statistic for user-defined windows
}
\examples{
RV_TDT(filepath.vcf, filepath.ped)
RV_TDT(filepath.vcf, filepath.ped, window.size=10, window.type="M")

}
