% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTransmittedRareVarCounts.R
\name{getTransmittedRareVarCounts}
\alias{getTransmittedRareVarCounts}
\title{Get counts of the number of times each rare variants is transmitted from parents to offspring}
\usage{
getTransmittedRareVarCounts(vcf, ped, cutoff = 0.01)
}
\arguments{
\item{vcf}{vcf file}

\item{ped}{data frame containing pedigree information for the VCF}

\item{cutoff}{The MAF cutoff to determine which variants are "rare" (default is 0.01), for which transmission counts will calculated.}
}
\value{
results data frame containing the names of the SNVs that are considered rare, the number of times the rare variant was transmitted from parents to the affected offspring in the dataset, and the genomic position of the rare SNPs.
}
\description{
`getTransmittedRareVarCounts()` returns a data frame containing the count of rare variants transmitted from the parents to the affected offspring.
}
\examples{
 system.file("extdata", "hg38.ped.txt", package = "rvtrio")
ped <- read.table(fp.ped,header=TRUE)
head(ped)

fp.vcf <- system.file("extdata", "hg38.vcf", package = "rvtrio")
hg.assembly <- "hg38"
vcf <- VariantAnnotation::readVcf(fp.vcf, hg.assembly)

#Get no. of rare variants that are transmitted from parents to offspring
transmitted.rare.var.ct <- rvtrio::getTransmittedRareVarCounts(vcf, ped)

}
